@@ -5,7 +5,7 @@ homepage: SinaMN75.com
 publish_to: 'none'
 
 environment:
-  sdk: '>=2.17.0'
+  sdk: ">=2.18.6 <3.0.0"
   flutter: ">=3.0.0"
 
 dependencies:
@@ -16,32 +16,33 @@ dependencies:
 
   url_launcher: ^6.0.15
   get: ^4.6.3
-  package_info_plus: ^3.1.0
-  share_plus: ^6.3.2
-  file_picker: ^5.2.4
-  intl: ^0.18.0
+  package_info_plus: ^8.0.0
+  share_plus: ^9.0.0
+  file_picker: ^8.0.5
+  intl: ^0.19.0
   webviewx: ^0.2.1
   get_storage: ^2.0.3
   flutter_svg: ^2.0.5
   persian_tools: ^1.0.1
   group_button: ^5.2.0
-  image_cropper: ^3.0.0
+  image_cropper: ^4.0.1
   cached_network_image: ^3.2.1
   just_audio: ^0.9.27
   path_provider: ^2.0.11
-  flutter_local_notifications: ^14.0.0-dev.5
+  flutter_local_notifications: ^17.1.2
   flutter_easyloading: ^3.0.5
-  pin_code_fields: ^7.4.0
+  pin_code_fields: ^8.0.1
   intl_phone_number_input: ^0.7.0+2
-  flutter_image_compress: ^1.1.2
+  flutter_image_compress: ^2.3.0
   dio: ^5.1.1
-  dio_smart_retry: ^5.0.0
-  local_auth: ^2.1.6
-  local_auth_ios: ^1.0.12
+  dio_smart_retry: ^6.0.0
+  local_auth: ^2.2.0
+  # local_auth_ios: ^1.1.17
   video_thumbnail: ^0.5.3
   video_box: ^0.17.0
   crypto: ^3.0.2
   path_drawing: ^1.0.1
+  syncfusion_flutter_pdfviewer: ^20.4.54
 
 #  video_player: ^2.4.6
   #  local_auth_android: ^1.0.21
@@ -56,20 +57,20 @@ dependencies:
 #  image_picker: ^0.8.5+3
 #  fl_chart: ^0.62.0
 
-  syncfusion_localizations: ^21.1.41
-  syncfusion_flutter_datagrid_export: ^21.1.41
-  syncfusion_flutter_calendar: ^21.1.41
-  syncfusion_flutter_datepicker: ^21.1.41
-  syncfusion_flutter_charts: ^21.1.41
-  syncfusion_flutter_gauges: ^21.1.41
-  syncfusion_flutter_sliders: ^21.1.41
-  syncfusion_flutter_pdf: ^21.1.41
-  syncfusion_flutter_barcodes: ^21.1.41
-  syncfusion_officechart: ^21.1.41
-  syncfusion_flutter_maps: ^21.1.41
-  syncfusion_flutter_signaturepad: ^21.1.41
-  syncfusion_flutter_pdfviewer: ^21.1.41
-  syncfusion_flutter_treemap: ^21.1.41
+  # syncfusion_localizations: ^21.1.41
+  # syncfusion_flutter_datagrid_export: ^21.1.41
+  # syncfusion_flutter_calendar: ^21.1.41
+  # syncfusion_flutter_datepicker: ^21.1.41
+  # syncfusion_flutter_charts: ^21.1.41
+  # syncfusion_flutter_gauges: ^21.1.41
+  # syncfusion_flutter_sliders: ^21.1.41
+  # syncfusion_flutter_pdf: ^21.1.41
+  # syncfusion_flutter_barcodes: ^21.1.41
+  # syncfusion_officechart: ^22.1.34
+  # syncfusion_flutter_maps: ^21.1.41
+  # syncfusion_flutter_signaturepad: ^21.1.41
+  # syncfusion_flutter_pdfviewer: ^21.1.41
+  # syncfusion_flutter_treemap: ^21.1.41
 
 
 
