@@ -19,7 +19,8 @@ class FormBuilder extends StatefulWidget {
     this.uploadImageWidget,
     this.labelStyle,
     this.spaceBetween,
-    this.textFieldPadding = const EdgeInsets.symmetric(vertical: 12, horizontal: 20),
+    this.textFieldPadding =
+        const EdgeInsets.symmetric(vertical: 12, horizontal: 20),
     this.requiredText = "* This field is required",
     this.childrenText = "Children: ",
     this.crossAxisAlignment = CrossAxisAlignment.start,
@@ -64,7 +65,8 @@ class _FormBuilderState extends State<FormBuilder> {
   List<FormReadDto> forms = <FormReadDto>[];
 
   @override
-  Widget build(final BuildContext context) => _itemSwitcher(items: widget.formFields, isChildren: false);
+  Widget build(final BuildContext context) =>
+      _itemSwitcher(items: widget.formFields, isChildren: false);
 
   Widget _itemSwitcher({
     required final List<FormFieldReadDto> items,
@@ -77,27 +79,40 @@ class _FormBuilderState extends State<FormBuilder> {
         children: items.map((final FormFieldReadDto field) {
           switch (field.type) {
             case 0:
-              return _textField(field: field, isChildren: isChildren).marginSymmetric(vertical: widget.spaceBetween ?? 4);
+              return _textField(field: field, isChildren: isChildren)
+                  .marginSymmetric(vertical: widget.spaceBetween ?? 4);
             case 1:
-              return _textField(field: field, maxLine: 5, isChildren: isChildren).marginSymmetric(vertical: widget.spaceBetween ?? 4);
+              return _textField(
+                      field: field, maxLine: 5, isChildren: isChildren)
+                  .marginSymmetric(vertical: widget.spaceBetween ?? 4);
             case 2:
-              return _selectDialog(field: field, isChildren: isChildren).marginSymmetric(vertical: widget.spaceBetween ?? 4);
+              return _selectDialog(field: field, isChildren: isChildren)
+                  .marginSymmetric(vertical: widget.spaceBetween ?? 4);
             case 3:
-              return _selectDialog(field: field, isChildren: isChildren).marginSymmetric(vertical: widget.spaceBetween ?? 4);
+              return _selectDialog(field: field, isChildren: isChildren)
+                  .marginSymmetric(vertical: widget.spaceBetween ?? 4);
             case 4:
-              return _radio(field: field, isChildren: isChildren).marginSymmetric(vertical: widget.spaceBetween ?? 4);
+              return _radio(field: field, isChildren: isChildren)
+                  .marginSymmetric(vertical: widget.spaceBetween ?? 4);
             case 5:
-              return _textField(field: field, isChildren: isChildren).marginSymmetric(vertical: widget.spaceBetween ?? 4);
+              return _textField(field: field, isChildren: isChildren)
+                  .marginSymmetric(vertical: widget.spaceBetween ?? 4);
             case 6:
-              return _filePicker(field: field, isChildren: isChildren).marginSymmetric(vertical: widget.spaceBetween ?? 4);
+              return _filePicker(field: field, isChildren: isChildren)
+                  .marginSymmetric(vertical: widget.spaceBetween ?? 4);
             case 7:
-              return _imagePicker(field: field, isChildren: isChildren).marginSymmetric(vertical: widget.spaceBetween ?? 4);
+              return _imagePicker(field: field, isChildren: isChildren)
+                  .marginSymmetric(vertical: widget.spaceBetween ?? 4);
             case 8:
-              return _textField(field: field, isChildren: isChildren).marginSymmetric(vertical: widget.spaceBetween ?? 4);
+              return _textField(field: field, isChildren: isChildren)
+                  .marginSymmetric(vertical: widget.spaceBetween ?? 4);
             case 9:
-              return _textField(field: field, isChildren: isChildren).marginSymmetric(vertical: widget.spaceBetween ?? 4);
+              return _textField(field: field, isChildren: isChildren)
+                  .marginSymmetric(vertical: widget.spaceBetween ?? 4);
             case 10:
-              return _textField(field: field, isChildren: isChildren, obscureText: true).marginSymmetric(vertical: widget.spaceBetween ?? 4);
+              return _textField(
+                      field: field, isChildren: isChildren, obscureText: true)
+                  .marginSymmetric(vertical: widget.spaceBetween ?? 4);
             case 11:
               return _dateTimePicker(field: field, isChildren: isChildren);
             case 12:
@@ -110,8 +125,11 @@ class _FormBuilderState extends State<FormBuilder> {
         }).toList(),
       );
 
-  Widget _dateTimePicker({required final FormFieldReadDto field, final bool isChildren = true}) {
-    List<FormReadDto> children = forms.isNotEmpty ? forms.singleWhere((e) => e.id == field.id).children : [];
+  Widget _dateTimePicker(
+      {required final FormFieldReadDto field, final bool isChildren = true}) {
+    List<FormReadDto> children = forms.isNotEmpty
+        ? forms.singleWhere((e) => e.id == field.id).children
+        : [];
     String dateText = widget.selectDateText ?? "Select Date";
     String timeText = widget.selectTimeText ?? "Select Time";
     String dateTimeText = widget.selectTimeText ?? "Select Date Time";
@@ -136,7 +154,8 @@ class _FormBuilderState extends State<FormBuilder> {
                   : field.type == 12
                       ? timeText
                       : dateTimeText,
-              style: widget.labelStyle?.copyWith(color: widget.selectDateTimeTextColor),
+              style: widget.labelStyle
+                  ?.copyWith(color: widget.selectDateTimeTextColor),
             ),
           ),
         ).onTap(() async {
@@ -147,14 +166,17 @@ class _FormBuilderState extends State<FormBuilder> {
               firstDate: DateTime.now(),
               lastDate: DateTime(DateTime.now().year + 5),
             );
-            setter(() => dateText = "${date!.day} / ${date!.month} / ${date!.year}");
+            setter(() =>
+                dateText = "${date!.day} / ${date!.month} / ${date!.year}");
             if (isChildren) {
               children.removeWhere((e) => e.id == field.id);
-              children.add(FormReadDto(id: field.id, title: dateText, formField: field));
+              children.add(
+                  FormReadDto(id: field.id, title: dateText, formField: field));
               forms.singleWhere((e) => e.id == field.id).children = children;
             } else {
               forms.removeWhere((final FormReadDto e) => e.id == field.id);
-              forms.add(FormReadDto(id: field.id, title: dateText, formField: field));
+              forms.add(
+                  FormReadDto(id: field.id, title: dateText, formField: field));
             }
             widget.onFormChanged(forms);
           } else if (field.type == 12) {
@@ -165,11 +187,13 @@ class _FormBuilderState extends State<FormBuilder> {
             setter(() => timeText = "${time!.minute} : ${time!.hour}");
             if (isChildren) {
               children.removeWhere((e) => e.id == field.id);
-              children.add(FormReadDto(id: field.id, title: timeText, formField: field));
+              children.add(
+                  FormReadDto(id: field.id, title: timeText, formField: field));
               forms.singleWhere((e) => e.id == field.id).children = children;
             } else {
               forms.removeWhere((final FormReadDto e) => e.id == field.id);
-              forms.add(FormReadDto(id: field.id, title: timeText, formField: field));
+              forms.add(
+                  FormReadDto(id: field.id, title: timeText, formField: field));
             }
             widget.onFormChanged(forms);
           }
@@ -189,13 +213,18 @@ class _FormBuilderState extends State<FormBuilder> {
         leading: Text(field.label!, style: widget.labelStyle),
         trailing: selectedFile == File("")
             ? GestureDetector(
-                onTap: () => showFilePicker(action: (final List<File> files) => selectedFile = files.first),
-                child: widget.uploadFileWidget ?? const Icon(Icons.add_circle, size: 30),
+                onTap: () => showFilePicker(
+                    action: (final List<File> files) =>
+                        selectedFile = files.first),
+                child: widget.uploadFileWidget ??
+                    const Icon(Icons.add_circle, size: 30),
               )
             : Stack(
                 children: <Widget>[
-                  widget.selectedFileWidget ?? const Icon(Icons.file_copy, size: 30),
-                  const Icon(Icons.clear, size: 15).onTap(() => setter(() => selectedFile = File(""))),
+                  widget.selectedFileWidget ??
+                      const Icon(Icons.file_copy, size: 30),
+                  const Icon(Icons.clear, size: 15)
+                      .onTap(() => setter(() => selectedFile = File(""))),
                 ],
               ),
       ),
@@ -213,13 +242,18 @@ class _FormBuilderState extends State<FormBuilder> {
         leading: Text(field.label!, style: widget.labelStyle),
         trailing: selectedFile == File("")
             ? GestureDetector(
-                onTap: () => showFilePicker(action: (final List<File> files) => selectedFile = files.first),
-                child: widget.uploadImageWidget ?? const Icon(Icons.add_circle, size: 30),
+                onTap: () => showFilePicker(
+                    action: (final List<File> files) =>
+                        selectedFile = files.first),
+                child: widget.uploadImageWidget ??
+                    const Icon(Icons.add_circle, size: 30),
               )
             : Stack(
                 children: <Widget>[
-                  imageFile(selectedFile, width: 70, height: 70, borderRadius: 7),
-                  const Icon(Icons.clear, size: 15).onTap(() => setter(() => selectedFile = File(""))),
+                  imageFile(selectedFile,
+                      width: 70, height: 70, borderRadius: 7),
+                  const Icon(Icons.clear, size: 15)
+                      .onTap(() => setter(() => selectedFile = File(""))),
                 ],
               ),
       ),
@@ -230,7 +264,9 @@ class _FormBuilderState extends State<FormBuilder> {
     required final FormFieldReadDto field,
     final bool isChildren = true,
   }) {
-    List<FormReadDto> children = forms.isNotEmpty ? forms.singleWhere((e) => e.id == field.id).children : [];
+    List<FormReadDto> children = forms.isNotEmpty
+        ? forms.singleWhere((e) => e.id == field.id).children
+        : [];
     final RxBool radioValue = false.obs;
     return Obx(
       () => iconTextHorizontal(
@@ -242,19 +278,26 @@ class _FormBuilderState extends State<FormBuilder> {
             radioValue.value = value!;
             if (radioValue.value) {
               if (isChildren) {
-                forms.singleWhere((element) => element.id == field.id).children.removeWhere((e) => e.id == field.id);
-                children.add(FormReadDto(id: field.id, title: "true", formField: field));
+                forms
+                    .singleWhere((element) => element.id == field.id)
+                    .children
+                    .removeWhere((e) => e.id == field.id);
+                children.add(
+                    FormReadDto(id: field.id, title: "true", formField: field));
               } else {
                 forms.removeWhere((final FormReadDto e) => e.id == field.id);
-                forms.add(FormReadDto(id: field.id, title: "true", formField: field));
+                forms.add(
+                    FormReadDto(id: field.id, title: "true", formField: field));
               }
             } else {
               if (isChildren) {
                 children.removeWhere((e) => e.id == field.id);
-                children.add(FormReadDto(id: field.id, title: "false", formField: field));
+                children.add(FormReadDto(
+                    id: field.id, title: "false", formField: field));
               } else {
                 forms.removeWhere((final FormReadDto e) => e.id == field.id);
-                forms.add(FormReadDto(id: field.id, title: "false", formField: field));
+                forms.add(FormReadDto(
+                    id: field.id, title: "false", formField: field));
               }
             }
           },
@@ -269,7 +312,9 @@ class _FormBuilderState extends State<FormBuilder> {
   }) {
     final List<String> items = field.optionList!.split(",");
     final List<String> selectedItems = <String>[];
-    List<FormReadDto> children = forms.isNotEmpty ? forms.singleWhere((e) => e.id == field.id).children : [];
+    List<FormReadDto> children = forms.isNotEmpty
+        ? forms.singleWhere((e) => e.id == field.id).children
+        : [];
     String result = "";
     return iconTextVertical(
       leading: Text(field.label!, style: widget.labelStyle),
@@ -285,8 +330,10 @@ class _FormBuilderState extends State<FormBuilder> {
                 margin: MediaQuery.of(context).viewInsets,
                 height: 300,
                 decoration: BoxDecoration(
-                  color: context.theme.backgroundColor.withOpacity(0.9),
-                  borderRadius: const BorderRadius.only(topLeft: Radius.circular(20), topRight: Radius.circular(20)),
+                  color: context.theme.colorScheme.surface.withOpacity(0.9),
+                  borderRadius: const BorderRadius.only(
+                      topLeft: Radius.circular(20),
+                      topRight: Radius.circular(20)),
                 ),
                 padding: const EdgeInsets.all(20),
                 child: ListView.builder(
@@ -306,25 +353,34 @@ class _FormBuilderState extends State<FormBuilder> {
                       result = selectedItems.join(",");
                       if (selectedItems.isNotEmpty) {
                         if (isChildren) {
-                          children.removeWhere((final FormReadDto e) => e.id == field.id);
-                          children.add(FormReadDto(id: field.id, title: result, formField: field));
+                          children.removeWhere(
+                              (final FormReadDto e) => e.id == field.id);
+                          children.add(FormReadDto(
+                              id: field.id, title: result, formField: field));
                         } else {
-                          forms.removeWhere((final FormReadDto e) => e.id == field.id);
-                          forms.add(FormReadDto(id: field.id, title: result, formField: field));
+                          forms.removeWhere(
+                              (final FormReadDto e) => e.id == field.id);
+                          forms.add(FormReadDto(
+                              id: field.id, title: result, formField: field));
                         }
                       } else {
                         if (isChildren) {
-                          children.removeWhere((final FormReadDto e) => e.id == field.id);
+                          children.removeWhere(
+                              (final FormReadDto e) => e.id == field.id);
                         } else {
-                          forms.removeWhere((final FormReadDto e) => e.id == field.id);
+                          forms.removeWhere(
+                              (final FormReadDto e) => e.id == field.id);
                         }
                       }
-                      forms.singleWhere((e) => e.id == field.id).children = children;
+                      forms.singleWhere((e) => e.id == field.id).children =
+                          children;
                       // finish
                       widget.onFormChanged(forms);
                     },
                     child: Text(items[index], style: widget.labelStyle),
-                  ).alignAtCenter().marginSymmetric(vertical: 10, horizontal: 15),
+                  )
+                      .alignAtCenter()
+                      .marginSymmetric(vertical: 10, horizontal: 15),
                 ),
               ),
             ),
@@ -339,7 +395,10 @@ class _FormBuilderState extends State<FormBuilder> {
                     border: Border.all(color: context.theme.primaryColor),
                   ),
               child: iconTextHorizontal(
-                leading: Text(widget.selectDialogText ?? "Select item", style: widget.labelStyle).alignAtCenterLeft().expanded(),
+                leading: Text(widget.selectDialogText ?? "Select item",
+                        style: widget.labelStyle)
+                    .alignAtCenterLeft()
+                    .expanded(),
                 trailing: Icon(Icons.keyboard_arrow_down),
               ),
             ),
@@ -363,25 +422,33 @@ class _FormBuilderState extends State<FormBuilder> {
                 ),
               ).onTap(
                 () {
-                  if (selectedItems.isNotEmpty) setter(() => selectedItems.removeAt(index));
+                  if (selectedItems.isNotEmpty)
+                    setter(() => selectedItems.removeAt(index));
                   // update forms:
                   result = selectedItems.join(",");
                   if (selectedItems.isNotEmpty) {
                     if (isChildren) {
-                      children.removeWhere((final FormReadDto e) => e.id == field.id);
-                      children.add(FormReadDto(id: field.id, title: result, formField: field));
+                      children.removeWhere(
+                          (final FormReadDto e) => e.id == field.id);
+                      children.add(FormReadDto(
+                          id: field.id, title: result, formField: field));
                     } else {
-                      forms.removeWhere((final FormReadDto e) => e.id == field.id);
-                      forms.add(FormReadDto(id: field.id, title: result, formField: field));
+                      forms.removeWhere(
+                          (final FormReadDto e) => e.id == field.id);
+                      forms.add(FormReadDto(
+                          id: field.id, title: result, formField: field));
                     }
                   } else {
                     if (isChildren) {
-                      children.removeWhere((final FormReadDto e) => e.id == field.id);
+                      children.removeWhere(
+                          (final FormReadDto e) => e.id == field.id);
                     } else {
-                      forms.removeWhere((final FormReadDto e) => e.id == field.id);
+                      forms.removeWhere(
+                          (final FormReadDto e) => e.id == field.id);
                     }
                   }
-                  forms.singleWhere((e) => e.id == field.id).children = children;
+                  forms.singleWhere((e) => e.id == field.id).children =
+                      children;
                   // finish
                   widget.onFormChanged(forms);
                 },
@@ -397,7 +464,9 @@ class _FormBuilderState extends State<FormBuilder> {
     required final FormFieldReadDto field,
     final bool isChildren = true,
   }) {
-    List<FormReadDto> children = forms.isNotEmpty ? forms.singleWhere((e) => e.id == field.id).children : [];
+    List<FormReadDto> children = forms.isNotEmpty
+        ? forms.singleWhere((e) => e.id == field.id).children
+        : [];
     final GroupButtonController controller = GroupButtonController();
     final List<String> items = field.optionList!.split(",");
     final List<String> selectedItems = <String>[];
@@ -408,7 +477,8 @@ class _FormBuilderState extends State<FormBuilder> {
         trailing: GroupButton<String>(
           controller: controller,
           enableDeselect: true,
-          onSelected: (final String value, final int index, final bool isSelected) {
+          onSelected:
+              (final String value, final int index, final bool isSelected) {
             if (field.type! == 3) {
               if (isSelected) {
                 selectedItems.clear();
@@ -426,10 +496,12 @@ class _FormBuilderState extends State<FormBuilder> {
             if (selectedItems.isNotEmpty) {
               if (isChildren) {
                 children.removeWhere((final FormReadDto e) => e.id == field.id);
-                children.add(FormReadDto(id: field.id, title: result, formField: field));
+                children.add(
+                    FormReadDto(id: field.id, title: result, formField: field));
               } else {
                 forms.removeWhere((final FormReadDto e) => e.id == field.id);
-                forms.add(FormReadDto(id: field.id, title: result, formField: field));
+                forms.add(
+                    FormReadDto(id: field.id, title: result, formField: field));
               }
             } else {
               if (isChildren) {
@@ -453,7 +525,9 @@ class _FormBuilderState extends State<FormBuilder> {
     final bool isChildren = true,
     final bool obscureText = false,
   }) {
-    List<FormReadDto> children = forms.isNotEmpty ? forms.singleWhere((e) => e.id == field.id).children : [];
+    List<FormReadDto> children = forms.isNotEmpty
+        ? forms.singleWhere((e) => e.id == field.id).children
+        : [];
     return Column(
       crossAxisAlignment: widget.crossAxisAlignment,
       children: <Widget>[
@@ -470,17 +544,23 @@ class _FormBuilderState extends State<FormBuilder> {
             onChanged: (final String value) {
               if (value != "") {
                 if (isChildren) {
-                  if (children.isNotEmpty) children.removeWhere((final FormReadDto e) => e.id == field.id);
-                  children.add(FormReadDto(id: field.id, title: value, formField: field));
+                  if (children.isNotEmpty)
+                    children
+                        .removeWhere((final FormReadDto e) => e.id == field.id);
+                  children.add(FormReadDto(
+                      id: field.id, title: value, formField: field));
                 } else {
                   forms.removeWhere((final FormReadDto e) => e.id == field.id);
-                  forms.add(FormReadDto(id: field.id, title: value, formField: field));
+                  forms.add(FormReadDto(
+                      id: field.id, title: value, formField: field));
                 }
                 forms.singleWhere((e) => e.id == field.id).children = children;
                 widget.onFormChanged(forms);
               } else {
                 if (isChildren) {
-                  if (children.isNotEmpty) children.removeWhere((final FormReadDto e) => e.id == field.id);
+                  if (children.isNotEmpty)
+                    children
+                        .removeWhere((final FormReadDto e) => e.id == field.id);
                 } else {
                   forms.removeWhere((final FormReadDto e) => e.id == field.id);
                 }
@@ -494,7 +574,8 @@ class _FormBuilderState extends State<FormBuilder> {
             decoration: InputDecoration(
               counter: const SizedBox(),
               fillColor: widget.textFieldFillColor,
-              hintStyle: context.textTheme.subtitle2!.copyWith(color: context.theme.hintColor),
+              hintStyle: context.textTheme.titleSmall!
+                  .copyWith(color: context.theme.hintColor),
               filled: true,
               contentPadding: widget.textFieldPadding,
               focusedBorder: widget.focusedBorder,
@@ -520,7 +601,9 @@ class _FormBuilderState extends State<FormBuilder> {
     );
   }
 
-  void validateForm({required final GlobalKey<FormState> key, required final VoidCallback action}) {
+  void validateForm(
+      {required final GlobalKey<FormState> key,
+      required final VoidCallback action}) {
     if (key.currentState!.validate()) action();
   }
 
