@@ -1,5 +1,7 @@
 import 'package:share_plus/share_plus.dart';
 
-void shareText(final String text, {final String? subject}) => Share.share(text, subject: subject);
+void shareText(final String text, {final String? subject}) =>
+    Share.share(text, subject: subject);
 
-void shareFile(final List<String> file, final String text) => Share.shareFiles(file, text: text);
+void shareFile(final List<XFile> file, final String text) =>
+    Share.shareXFiles(file, text: text);
